requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==26.1.0
cffi==2.0.0
davey==0.1.4
discord.py==2.7.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
pycparser==3.0
PyNaCl==1.6.2
python-dotenv==1.2.2
typing_extensions==4.15.0
yarl==1.23.0
yt-dlp==2026.3.17
